requirements.txt
  fastapi
  uvicorn 
  python-multipart
  pillow 
  pymongo 
  cloudinary 
  pypdf 
  docx2txt
  python-dotenv




